requirements.txt
fastapi[standard]>=0.115.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
uvicorn>=0.30.0
python-dotenv